refactor(dashboard): consolidate duplicate utilities into shared libs
- Extract formatDuration(ms) from activity-swimlane into lib/format-time as formatDurationMs
- Remove compactCopy duplicates (agent-detail-state, agent-profile) and use existing trimText from lib/truncate
- Replace silent .catch(() => {}) with console.warn in mcp.ts initialized notification
